backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
opencv-python-headless>=4.8.1.78
numpy>=1.26.0
tensorflow>=2.16.0
scikit-learn>=1.3.2
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
pydantic>=2.5.0
pydantic-settings>=2.1.0
Pillow>=10.1.0
websockets>=12.0
scipy>=1.11.4
